Split up test into multiple files.
It was getting a little unwieldy
